move helper to separate file; makes page object use openapi types
